LayoutTests/imported/w3c/web-platform-tests/css/css-overflow/scroll-markers/scroll-marker-inert-001-expected.txt


PASS ::scroll-marker is inert because its ::scroll-marker-group is
